requirements/test.txt
coverage==4.5.4
isort==4.3.21
pycodestyle==2.5.0
pylint-django==2.0.11
mypy==0.790
